ADDED EVENT SYSTEM AND PRIMARY FORM OF EVENT MANAGER:
Some kind of event system has been implemented with a default event
manager. For now this cannot evolve until windows and application class
are introduced (next step). A lot of code has to be commented. That
activity will be achieved after the implementation of Application class
and window class with GLFW.